+++ fancyshowcaseview/src/test/java/me/toptas/fancyshowcase/PropertiesTest.kt
@@ -1,0 +1,14 @@
+package me.toptas.fancyshowcase
+
+import me.toptas.fancyshowcase.internal.Properties
+import org.junit.Test
+
+class PropertiesTest {
+
+    @Test
+    fun testInitialValues() {
+        val props = Properties()
+        assert(props.fancyId == null)
+    }
+
+}